detect_species_component/requirements.txt
attrs==21.2.0
azure-core==1.14.0
azure-storage-blob==12.8.1
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
decorator==5.0.9
distro==1.5.0
docopt==0.6.2
Flask==1.1.2
Flask-Cors>=3.0
idna==2.10
importlib-metadata==2.1.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
msrest==0.6.21
nbformat==5.1.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.5.2.52
packaging==20.9
pigar==0.10.0
Pillow==7.2.0
pipreqs==0.4.10
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-build==0.11.1
six==1.16.0
traitlets==4.3.3
urllib3==1.26.4
Werkzeug==1.0.1
yarg==0.1.9
zipp==1.2.0